requirements.txt
pandas==0.24.2
google-cloud-bigquery==1.17.0
psycopg2==2.8.3
